requirements.txt
Flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
google-generativeai==0.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.2.0
pytesseract==0.3.13
Pillow==10.2.0
reportlab==4.4.3
pandas==2.1.4
numpy==1.24.3
jinja2==3.1.3
werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
markdown==3.5.1 